# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores": {
      "get": {
        "operationId": "PublicController_stores",
        "parameters": [
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -90,
              "maximum": 90,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -180,
              "maximum": 180,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
...
    "/api/orders": {
      "post": {
        "operationId": "CustomerOrdersController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/admin/overview": {
      "get": {
        "operationId": "AdminController_overview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/kpis": {
      "get": {
        "operationId": "AdminController_kpisRange",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "grain",
            "required": false,
            "in": "query",
            "schema": {
              "default": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "default": "all",
              "type": "string",
              "enum": [
                "all",
                "customers",
                "merchants",
                "staff"
              ]
            }
          },
          {
            "name": "category_slug",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 60,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "active",
                "rejected",
                "suspended",
                "deleted"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": false,
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/staff": {
      "post": {
        "operationId": "AdminController_createStaff",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/profile": {
      "patch": {
        "operationId": "AdminController_updateProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/status": {
      "post": {
        "operationId": "AdminController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/role": {
      "post": {
        "operationId": "AdminController_setRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/wallet-grant": {
      "post": {
        "operationId": "AdminController_grantWallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletGrantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/account-requests": {
      "get": {
        "operationId": "AdminController_accountRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/account-requests/{id}/decide": {
      "post": {
        "operationId": "AdminController_decideAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/password-resets": {
      "get": {
        "operationId": "AdminController_passwordResets",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": false,
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/password-resets/{id}/decide": {
      "post": {
        "operationId": "AdminController_decideReset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/stores": {
      "get": {
        "operationId": "AdminController_stores",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": false,
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/stores/{id}/status": {
      "post": {
        "operationId": "AdminController_storeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/stores/{id}/commission": {
      "post": {
        "operationId": "AdminController_storeCommission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreCommissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/stores/{id}/category": {
      "post": {
        "operationId": "AdminController_storeCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/orders": {
      "get": {
        "operationId": "AdminController_orders",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          },
          {
            "name": "store_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5000,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/wallets/transactions": {
      "get": {
        "operationId": "AdminController_walletTx",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": false,
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/wallets/transactions/{id}/respond": {
      "post": {
        "operationId": "AdminController_respondWallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletRespondDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/settings": {
      "get": {
        "operationId": "AdminController_settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/settings/{key}": {
      "put": {
        "operationId": "AdminController_setSetting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/permissions/catalog": {
      "get": {
        "operationId": "AdminController_permissionCatalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wasl API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^(77|78|71|73)\\d{7}$"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 72
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "user_type": {
            "type": "string",
            "enum": [
              "customer",
              "merchant"
            ]
          },
          "business_name": {
            "type": "string",
            "maxLength": 120
          },
          "business_category_id": {
            "type": "string",
            "format": "uuid"
          },
          "city": {
            "type": "string",
            "maxLength": 80
          },
          "district": {
            "type": "string",
            "maxLength": 80
          },
          "address": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "phone",
          "password",
          "name",
          "user_type"
        ]
      },
...
      "DecideDto": {
        "type": "object",
        "properties": {
          "approve": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "approve"
        ]
      },
...
    }
  }
}